+++ cw-vault-standard/src/helper.rs
@@ -2,7 +2,7 @@ use std::marker::PhantomData;
 
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{
-    coin, to_binary, Addr, CosmosMsg, Deps, QuerierWrapper, StdResult, Uint128, WasmMsg,
+    coin, to_json_binary, Addr, CosmosMsg, Deps, QuerierWrapper, StdResult, Uint128, WasmMsg,
 };
 use schemars::JsonSchema;
 use serde::Serialize;
@@ -89,7 +89,7 @@ where
 
         Ok(WasmMsg::Execute {
             contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Deposit {
+            msg: to_json_binary(&VaultStandardExecuteMsg::<E>::Deposit {
                 amount: amount.clone(),
                 recipient,
             })?,
@@ -104,7 +104,7 @@ where
     pub fn deposit_cw20(&self, amount: Uint128, recipient: Option<String>) -> StdResult<CosmosMsg> {
         Ok(WasmMsg::Execute {
             contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Deposit { amount, recipient })?,
+            msg: to_json_binary(&VaultStandardExecuteMsg::<E>::Deposit { amount, recipient })?,
             funds: vec![],
         }
         .into())
@@ -119,7 +119,7 @@ where
         let amount = amount.into();
         Ok(WasmMsg::Execute {
             contract_addr: self.addr.to_string(),
-            msg: to_binary(&VaultStandardExecuteMsg::<E>::Redeem {
+            msg: to_json_binary(&VaultStandardExecuteMsg::<E>::Redeem {
                 amount: amount.clone(),
                 recipient,
             })?,